Issue 137 AGENTS AND VERBS views with data export
We have characters and verbs. Need corresponding exports for entities and technologies.
We need to be able to separate these based on whether they are in a game,
artwork or narrative (Publication_type is a field on Creative Work),
and on which situation they occur in (Title of Situation they are mentioned in).