Issue 156 Avoid taxonomy term id collision with node id
Separate entity types like taxonomy and node use same number serie
as, for them, unique id. When we expose id from both of them in
a data export they might be naturly be identical.

To avoid that is all creative work taxonomy term id padded with
the value 2000. All node id is below 10000.